Time: 5 ms (32.01%), Space: 46.4 MB (5.36%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -25,7 +25,7 @@ class Solution {
         }
         else 
         {
-            return ans[mid];
+            return (double)ans[mid];
         }
     }
 }
